move image uploading function to helperviewmodel
