+++ app/src/main/java/com/example/helpersapp/viewModel/HelperViewModel.kt
@@ -1,7 +1,14 @@
 package com.example.helpersapp.viewModel
 
+import android.net.Uri
+import android.util.Log
 import androidx.lifecycle.ViewModel
+import com.example.helpersapp.ui.components.createUsername
+import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.ktx.Firebase
+import com.google.firebase.storage.ktx.storage
+import java.util.UUID
 
 class HelperViewModel: ViewModel()  {
 
@@ -34,4 +41,33 @@ class HelperViewModel: ViewModel()  {
             }
     }
 
+    private val storageReference = Firebase.storage.reference
+    fun uploadProfilePicture(uri: Uri) {
+        val useremail = Firebase.auth.currentUser?.email
+        val username = createUsername(useremail ?: "")
+        val profilePictureRef = storageReference.child("$username/profile_picture.jpg")
+
+        profilePictureRef.putFile(uri)
+            .addOnSuccessListener {
+                Log.d("FirebaseStorage", "Profile picture upload success")
+            }
+            .addOnFailureListener { exception ->
+                Log.e("FirebaseStorage", "Error uploading profile picture", exception)
+            }
+    }
+
+    fun uploadCertificate(uri: Uri) {
+        val useremail = Firebase.auth.currentUser?.email
+        val username = createUsername(useremail ?: "")
+        val certificateId = UUID.randomUUID().toString()
+        val certificateRef = storageReference.child("$username/certificates/$certificateId.pdf")
+
+        certificateRef.putFile(uri)
+            .addOnSuccessListener {
+                Log.d("FirebaseStorage", "Certificate upload success")
+            }
+            .addOnFailureListener { exception ->
+                Log.e("FirebaseStorage", "Error uploading certificate", exception)
+            }
+    }
 }
